feat: 푸터 활성화 기능 추가 및 폴더 위치 수정
